+++ home.js
@@ -1,28 +1,46 @@
-import { COMMANDS } from './consts.js';
+import { COMMANDS, ROOT_COMMANDS } from './consts.js';
+import { getCommandWithArgs } from './utils.js';
 
-const script = `import('/commands.js')
-	.then(({ observeCommands }) => observeCommands());`;
+const script = `import { observeCommands, initRootCommands, registerRootCommand, registerCommand } from '/commands.js'
+registerRootCommand('--menu', () => document.getElementById('menu').showPopover());
+registerRootCommand('--help', () => document.getElementById('help').showPopover());
+registerRootCommand('--log', (event, ...args) => console.log({ event, args }));
+observeCommands();
+initRootCommands();
+`;
 
-const hash = await crypto.subtle.digest('SHA-384', new TextEncoder().encode(script))
+const style = `:root {
+	color-scheme: dark;
+
+	& body.foo {
+		filter: invert(1);
+	}
+}`;
+
+const sri = async text => await crypto.subtle.digest('SHA-384', new TextEncoder().encode(text))
 	.then(digest => new Uint8Array(digest).toBase64())
 	.then(hash => `sha384-${hash}`);
 
+const scriptSRI = await sri(script);
+const styleSRI = await sri(style);
+
 const headers = new Headers({
 	'Content-Type': 'text/html',
-	'Content-Security-Policy': `default-src 'self'; script-src 'self' '${hash}'; media-src https://0eff4f4c-7f45-405c-8cf6-f7a3b3c1f07e.mdnplay.dev;`,
+	'Content-Security-Policy': `default-src 'self'; script-src 'self' '${scriptSRI}'; style-src '${styleSRI}'; media-src https://0eff4f4c-7f45-405c-8cf6-f7a3b3c1f07e.mdnplay.dev;`,
 });
 
 
 const doc = `<!DOCTYPE html>
-<html>
+<html id="doc">
 	<head>
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width" />
 		<meta name="color-scheme" content="light dark" />
 		<script src="/node_modules/@shgysk8zer0/polyfills/browser.min.js" referrerpolicy="no-referrer" defer=""></script>
-		<script type="module" integrity="${hash}">${script}</script>
+		<script type="module" integrity="${scriptSRI}">${script}</script>
+		<style integrity="${styleSRI}" id="style">${style}</style>
 	</head>
-	<body>
+	<body id="top">
 		<header id="header"></header>
 		<nav id="nav">
 			<button type="button" command="${COMMANDS.disable}" commandfor="btn">Disable</button>
@@ -31,6 +49,14 @@ const doc = `<!DOCTYPE html>
 			<button type="button" command="${COMMANDS.showModal}" commandfor="dialog">Show Modal Dialog</button>
 			<button type="button" command="${COMMANDS.showPopover}" commandfor="popover">Show Popover</button>
 			<button type="button" command="${COMMANDS.togglePopover}" commandfor="popover">Toggle Popover</button>
+			<button type="button" command="${ROOT_COMMANDS.help}" commandfor="doc" accesskey="h">Help</button>
+			<button type="button" command="${COMMANDS.disable}" commandfor="style">Disable Styles</button>
+			<button type="button" command="${COMMANDS.enable}" commandfor="style">Enable Styles</button>
+			<button type="button" command="${getCommandWithArgs(COMMANDS.addClass, 'foo', 'bar')}:foo:bar" commandfor="top">Add <code>foo</code> <code>bar</code></button>
+			<button type="button" command="${getCommandWithArgs(COMMANDS.removeClass, 'foo', 'bar')}" commandfor="top">Remove <code>foo</code> <code>bar</code></button>
+			<button type="button" command="--menu" commandfor="doc" accesskey="m">Menu</button>
+			<button type="button" command="${COMMANDS.requestFullscreen}" commandfor="nav">Fullscreen</button>
+			<button type="button" command="${ROOT_COMMANDS.exitFullscreen}" commandfor="doc">Exit Fullscreen</button>
 		</nav>
 		<main id="main"></main>
 		<aside id="sidebar"></aside>
@@ -53,6 +79,15 @@ const doc = `<!DOCTYPE html>
 		<div id="popover" popover="auto">
 			<p>This is a test of the <code>popover</code> API.</p>
 			<button type="button" command="${COMMANDS.hidePopover}" commandfor="popover">Hide</button>
+		</div>
+		<menu id="menu" popover="auto">${[...Object.values(ROOT_COMMANDS), '--help'].map(command => `
+			<li><button type="button" command="${command}" commandfor="doc"><code>${command}</code></button></li>
+		`).join('')}
+			<li><button type="button" command="--log:foo:bar:abc:123" commandfor="doc">Log</button></li>
+		</menu>
+		<div id="help" popover="auto">
+			<p>Lorem Ipsum</p>
+			<button type="button" command="hide-popover" commandfor="help">Close</button>
 		</div>
 	</body>
 </html>`;